+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 
 # 2.3.1
 ### features
-- add restaurant breakfast checklist menu(hra-done)
+- add restaurant breakfast checklist menu (hra-done)
+- add order employee get customer (hra-done)
 ### bug fixes
 - move salons filters to url query (hra-done)
 - replace redirect link in table orders (hra-done)